Rework to single slider; merge PWM components
